execute command directly with acction server without a service
