+++ niryo_one_commander/scripts/niryo_one_commander_node.py
@@ -22,7 +22,7 @@ import rospy
 
 from position_manager import PositionManager
 from niryo_one_robot_state_publisher import NiryoRobotStatePublisher
-from robot_action_server import RobotActionServer
+#from robot_action_server import RobotActionServer
 from robot_commander import RobotCommander 
 
 class NiryoOneCommanderNode(): 
@@ -37,10 +37,7 @@ class NiryoOneCommanderNode():
         
         # robot commander 
         self.robot_commander = RobotCommander(self.pos_manager)
-        
-        # Load ROS action interface 
-        self.robot_action_server = RobotActionServer(self.pos_manager)
-        self.robot_action_server.start()
+        self.robot_commander.start()
         
 if __name__ == '__main__':
     rospy.init_node('niryo_one_commander_node')
