+++ niryo_one_commander/src/niryo_one_commander/validate_params.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# robot_action_server.py
+# validate_params.py
 # Copyright (C) 2017 Niryo
 # All rights reserved.
 #
@@ -17,179 +17,22 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import rospy, sys
-import actionlib
-import threading
+
+
+import rospy 
 from math import sqrt
 
-# Lib
-from niryo_one_commander.command_type import CommandType
-from niryo_one_commander.command_status import CommandStatus
 from niryo_one_commander.robot_commander_exception import RobotCommanderException
-from position_manager import PositionManager
-from robot_commander import RobotCommander
-# Messages
-from std_msgs.msg import Bool
-from niryo_one_msgs.msg import RobotMoveCommand
+from niryo_one_commander.command_status import CommandStatus
+from niryo_one_commander.command_type import CommandType
 
-# Services
-from std_srvs.srv import SetBool
-from niryo_one_msgs.srv import GetInt
-from niryo_one_msgs.srv import RobotMove
+class ValidateParameters(): 
 
-# Action msgs
-from niryo_one_msgs.msg import RobotMoveAction
-from niryo_one_msgs.msg import RobotMoveGoal
-from niryo_one_msgs.msg import RobotMoveResult
-#
-# This class is the main interface to use niryo_one_commander
-# It implements an action server that will :
-# - Check if no other command is being processed
-# - Validate params
-# - Execute command and return status + message
-# - Cancel command if asked
-#
-
-class RobotActionServer:
-
-    def __init__(self, position_manager):
-        
+    def __init__(self, validation, position_manager):
+        self.validation = validation 
         self.pos_manager = position_manager
-        self.robot_commander = RobotCommander(self.pos_manager)
-        self.server = actionlib.ActionServer('niryo_one/commander/robot_action',
-		 RobotMoveAction, self.on_goal, self.on_cancel, auto_start=False)
-    
-        self.current_goal_handle = None
 
-        self.learning_mode_on = False 
-        self.joystick_enabled = False
-
-        # Load validation params from rosparams
-        self.validation = rospy.get_param("/niryo_one/robot_command_validation")
-        
-        
-        self.is_active_server = rospy.Service(
-                'niryo_one/commander/is_active', GetInt, self.callback_is_active)
-
-        self.learning_mode_subscriber = rospy.Subscriber(
-                '/niryo_one/learning_mode', Bool, self.callback_learning_mode)
-        self.joystick_enabled_subscriber = rospy.Subscriber('/niryo_one/joystick_interface/is_enabled', 
-                Bool, self.callback_joystick_enabled)
-
-    def callback_learning_mode(self, msg):
-        activate = msg.data
-
-        if not self.learning_mode_on and activate:
-            self.cancel_current_command()
-
-        self.learning_mode_on = activate
-
-    def callback_joystick_enabled(self, msg):
-        self.joystick_enabled = msg.data
-
-    def callback_is_active(self, req):
-        if self.current_goal_handle is not None:
-            return 1
-        return 0
-
-    def start(self):
-        self.server.start()
-        rospy.loginfo("Action Server started (Robot Commander)")
-
-    def create_result(self, status, message):
-        result = RobotMoveResult()
-        result.status = status
-        result.message = message
-        return result
-
-    def on_goal(self, goal_handle):
-        rospy.loginfo("Robot Action Server - Received goal. Check if exists")
-
-        # Check if learning mode ON
-        if self.learning_mode_on:
-            result = self.create_result(CommandStatus.LEARNING_MODE_ON, "You need to deactivate learning mode to execute a new command")
-            goal_handle.set_rejected(result)
-            return
-
-        # Check if joystick enabled
-        if self.joystick_enabled:
-            result = self.create_result(CommandStatus.JOYSTICK_ENABLED, "You need to deactivate joystick to execute a new command")
-            goal_handle.set_rejected(result)
-            return
-
-        # check if still have a goal -> set_rejected() 
-        if self.current_goal_handle is not None:
-            result = self.create_result(CommandStatus.GOAL_STILL_ACTIVE, "Current command is still active. Cancel it if you want to execute a new one")
-            goal_handle.set_rejected(result)
-            return
-      
-        # validate parameters -> set_rejected (msg : validation or commander error)
-        try:
-            rospy.loginfo("Robot Acton Sever - checking paramter Validity")
-            self.validate_params(goal_handle.goal.goal.cmd)
-        except RobotCommanderException as e:
-            result = self.create_result(e.status, e.message)
-            goal_handle.set_rejected(result)
-            rospy.loginfo("Robot Action Server - Invalid parameters")
-            return
-        
-        # set accepted
-        self.current_goal_handle = goal_handle
-        self.current_goal_handle.set_accepted()
-        rospy.loginfo("Robot Action Server - Goal has been accepted")
-
-        # Launch compute + execution in a new thread
-        w = threading.Thread(name="worker", target=self.execute_command)
-        w.start()
-        rospy.loginfo("Robot Action Server : Executing command in a new thread")
-
-    def on_cancel(self, goal_handle):
-        rospy.loginfo("Received cancel command")
-
-        if goal_handle == self.current_goal_handle:
-            self.cancel_current_command()
-        else:
-            rospy.loginfo("Robot Action Server - No current goal, nothing to do")
-
-    def execute_command(self):
-        cmd = self.current_goal_handle.goal.goal.cmd 
-        (status, message) = self.robot_commander.execute_command(cmd)
-        result = self.create_result(CommandStatus.ROS_ERROR, "error with executing commad")
-        response = None
-        try: 
-            (status, message) = self.robot_commander.execute_command(cmd)
-            response = self.create_result(status, message)
-            result = response 
-        except RobotCommanderException :
-            rospy.loginfo ("error with executing command ") 
-
-
-
-        if not response:
-            self.current_goal_handle.set_aborted(result)
-            rospy.loginfo("Execution has been aborted")
-        elif response.status == CommandStatus.SUCCESS:
-            self.current_goal_handle.set_succeeded(result)
-            rospy.loginfo("Goal has been set as succeeded")
-        elif response.status == CommandStatus.STOPPED:
-            self.current_goal_handle.set_canceled(result)
-            rospy.loginfo("Goal has been successfully canceled")
-        else:
-            self.current_goal_handle.set_aborted(result)
-            rospy.loginfo("Unknown result, goal has been set as aborted")
-        self.current_goal_handle = None
-
-    
-    # Send a cancel signal to Moveit interface
-    def cancel_current_command(self):
-        rospy.wait_for_service('niryo_one/commander/stop_command')
-        try:
-            cancel_command = rospy.ServiceProxy('niryo_one/commander/stop_command', SetBool)
-            cancel_command(True)
-        except rospy.ServiceException as e:
-            rospy.logwarn("Could not cancel current command : " + str(e))
-    
-    def validate_params(self, cmd):
+    def validate_params(self, cmd): 
         cmd_type = cmd.cmd_type
         if cmd_type == CommandType.JOINTS:
             self.validate_joints(cmd.joints)
